reorg: merge code from call.cpp into stack.cpp
